Rename _helpers.py to _utils.py and assert deleted entities in tests
- Rename mlflow/store/fs2db/_helpers.py -> _utils.py
- Assert that deleted experiments and runs exist in test data to ensure
  lifecycle_stage comparisons are not vacuously passing
